+++ src/bit-id/index.js
@@ -1,0 +1,5 @@
+import findLatestVersion from './find-latest-version';
+import parseBitFullId from './parse-bit-full-id';
+import parseBitInlineId from './parse-bit-inline-id';
+
+export { findLatestVersion, parseBitFullId, parseBitInlineId };